requirements.txt
flask==2.0.3
flask-cors==3.0.10
requests==2.26.0
websockets==10.1
psycopg2-binary==2.9.3
pandas==1.3.5
jupyter==1.0.0
matplotlib==3.5.1
folium==0.12.1
scikit-learn==1.0.2